core/device: create a dependency between a loop device and its backing file
When loop mounts are created via entries in /etc/fstab,
a RequiresMountsFor dependency on the backing file
is created.  However when loop mounts are created manually,
this dependency is not created.
This case cause unmounts to happen in the wrong order at shutdown.

So create a dependency by having udev set SYSTEMD_REQUIRES_MOUNTS_FOR,
and having the device unit read and process this.

Unlike Wants dependencies, we need to add this dependency to all aliases
so that the After ordering is effective.