core,systemctl: add bus API to retrieve processes of a unit
This adds a new GetProcesses() bus call to the Unit object which returns an
array consisting of all PIDs, their process names, as well as their full cgroup
paths. This is then used by "systemctl status" to show the per-unit process
tree.

This has the benefit that the client-side no longer needs to access the
cgroupfs directly to show the process tree of a unit. Instead, it now uses this
new API, which means it also works if -H or -M are used correctly, as the
information from the specific host is used, and not the one from the local
system.

Fixes: #2945
(cherry picked from commit 291d565a04263452c03beaf537773ade4f0b1617)

[mkoutny: drop(!) main part of the patch,
          take unit_{main,control}_pid functions only]
[mkoutny: dependency for bsc#1117382]