cgroup: Support 0-value for memory protection directives
These make sense to be explicitly set at 0 (which has a different effect
than the default, since it can affect processing of `DefaultMemoryXXX`).

Without this, it's not easily possible to relinquish memory protection
for a subtree, which is not great.

(cherry picked from commit 22bf131be278b95a4a204514d37a4344cf6365c6)

[mkoutny: implemented inside range condition, we don't have 681ae88e06 ("dbus-cgroup: simplify bus_cgroup_set_property()") v237~157^2]
[mkoutny: fixes bsc#1167471]