cgroup: Make empty assignments reset to default
When MemoryLow= or MemoryMin= is set, it is interpretted as setting the
values to infinity. This is inconsistent with the default initialization
to 0.
It'd be nice to interpret the empty assignment as fallback to
DefaultMemory* of parent slice, however, current DBus API cannot convey
such a NULL value, so stick to simply interpretting that as hard-wired
default.

(cherry picked from commit db2b8d2e2895010f3443a589c9c1f1dfb25256a6)

[mkoutny: fixes bsc#1167471]
[mkoutny: adjusted to v234 dbus properties setting
          dropped parts for MemoryMin,DefaultMemory{Min,Low} which we don't have]