pid1: No killing spree on switch_root from initrd
Some initrd context here:
- initrd is volatile filesystem (no persistent state once real root is
  available),
- initrd is `rootfs` type of filesystem, which means it cannot be
  pivot_root(2)'d,
  - therefore it is just hidden under MS_MOVE'd real root mount and its
    content is deleted (`rm -rf`) to release its memory footprint,
  - (even if it could pivot, switch_root unmounts it lazily with
     MNT_DETACH)
- rootfs is non-persistent single-use filesystem.

Historical excursion:
- initially no processes were killed on switch_root from initrd
- bd3fa1d2434a ("core: go on a killing spree when transitioning from initrd to main system") v187~107
- then they're wiped out with SIGTERM,SIGKILL combo,
- cee530bb23b7 ("switch-root: don't wait for processes") v187~39
- this was later relaxed with SIGTERM and switch_root doesn't wait for
  processes to terminate.

The initrd to real root transition is utilizing initrd-cleanup.service
that isolates the initrd-cleanup.target. That means all unnecessary
units are terminated with their stop commands (or forcefully [1]) and
there is no technical reason to initiate the killing spree before
switch_root from initrd.

There may be some other processes besides PID 1 remaining [2], they may
suffer from the switch_root but it's up to them to deal with the
transformed environment and it makes no harm to the rest of the system
[3].

If there was a misbehaving process in initrd environment, mere SIGTERM
is not guaranteed to resolve such situation (especially if its unit
failed to stop it during normal deactivation.)

If this killing spree was of any use (resource release, enforcing known
state), one could ask why to invoke it only during switch_root from initrd
and not during normal operation too? (E.g. during any isolate
transaction or cron.)

OK, the last question is why to change this now when it's been working
all the time. It gives more freedom to processes specified in [2] (as
said above, it makes no harm to allow more freedom) and it ensures
fail-safe behavior (for current SIGTERM victims to finish their job).

[1] Unless the units specify KillMode=none, SendSIGKILL=no or similar
    exemption from proper cleanup.
[2] E.g. root storage daemon, units configured to avoid full cleanup as
    above or processes external to PID 1 forking tree (kernel threads, UMHs).
[3] To reiterate, this focuses on switch_root from initrd at boot.
    During shutdown we need to properly sync and unmount the main
    (persistent) root filesystem and terminate any process that could
    preventing unmount.