WIP unit: Try loading rtemplate fragment
Rtemplate handle service exist as a normal service, it is distinguished
by a special unit file -- similar to ordinary templates whose unit
filename ends with '@' -- marked with '#' at the end of name.
When loading a unit, check existence of special fragment file denoting a
rtemplate unit.

(WIP name processing may not be complete)