job: Implement restart jobs for rtemplate handle services
Rtemplate restart job is similar to regular restart job. The speciality
is that it is effectively performed on the unit that is the current
generation of the handle service. Between stop and start halves of the
job, we switch the current generation of the rtemplate handle service,
thus stop and start operations are forwarded to different units.