core: Correct GC propagation via UnitRefs
Unit A (source) can point via a UnitRef to a unit B (target).
The UnitRef is supposed to pin B when A is pinned too.

Commit 4892084f09 ("manager: reduce complexity of unit_gc_sweep
(#3507)") introduced a shortcut by recursive propagation of GOOD marker.
However, it doesn't take into account UnitRef references (partially
covered in the commit 2641f02e23 ("pid1: fix collection of cycles of units which reference one another").

Consider scenario when:
- A would be GOOD
- B is processed before A
- B will be marked as UNSURE
- A will be marked as GOOD

B is collected (UnitRef is zeroed) despite it was referenced by A.

We cannot traverse forward edges in unit_gc_mark_good() (we only track
refs_by_target) so we should leverage the reverse marking introduced in
the commit 2641f02e23.
It even provisions for the drafted scenario by re-queueing unit B.
However, GC queue is not FIFO and UNSURE unit would not be processed
repeatedly.

Make GC queue FIFO (from LIFO) and re-process UNSURE units so that
UnitRef properly propagate GOOD markers forward.

FIXME The GC may not terminate if there is a graph loop consisting of
UnitRefs only.