unit: Add PropagatesRawStopTo= dependency
The relationship between rtemplate handle service and a generation
service should be such that stop job of the handle service induces a
stop job of all matured generation services. At the same time, restart
of the handle service should ignore matured generations.
This used to be captured by the PropagatesStopTo= dependency until
commit 017a7ba4f4 ("core: propagate stop too if restart is issued").

Since restart of a rtemplate service is not mere stop+start, introduce a
new dependency that distinguishes between stop job induction from a
restart job and a stop job. (That is actually what dependency atom
UNIT_ATOM_PROPAGATE_STOP, so we only expose it via a full dependency.)

This dependency is not meant to be configured manually at the moment, it
will only be implicitly created between rtemplate handle service and
its matured generations.