cgroup: Add simple support for misc cgroup controller
misc controller is a catch-all controller for various resources whose
list depends on kernel compile configuration.
Add the controller to the list of recognized controllers, so that users
can delegate it to their units.
The users can then apply whatever misc properties directly to the
cgroupfs, while systemd is kept oblivious of available properties the
kernel supports.

Ref: #30383