timer: Make timer_calculate_elapse more functional
The rebased monotonic timestamp relies on actual clock of the running
system (boils down to now(CLOCK_MONOTONIC)) but we want to calculate the
boottime with the explicitly provided triple_timestamp reference.

This is no functional change in timer code (beceause of
triple_timestamp_now(&ts)) but it enables the testing code to inject an
own reference triple_timestamp as convenient.